requirements.txt
streamlit
sentence-transformers
faiss-cpu
pdfplumber
python-docx
pandas
python-dotenv
langchain
python-dotenv
pypdf
huggingface-hub
PyMuPDF
langchain_community
langchain_text_splitters
langchain_huggingface
docx2txt
langchain_groq 
chromadb